Increased responsiveness of service.
Increased responsiveness of service by switching to asynchrounous HTTP
server and multithreading network requests.

Saw an improvement reducing mean reasponse time by approximately 60%:

From:

https://travis-ci.org/is-aws/is-aws-api/builds/77554545

To:

https://travis-ci.org/is-aws/is-aws-api/builds/78758030

Note: Bug in the version (2.4.11) of specs2 that Unfiltered depends on
prevents use of the `throwA[E].await` method. This is a workound
introducing the `Eventually` implicit transformation of `Future`
instances.